refactor(rules-builder): move openBuilder into ItemRulesTab factory
Pass the RulesBuilderWindow component in as a getter so the dialog
launcher stays out of the .svelte file. The factory is now fully
unit-testable without component context.
